pyproject.toml
[tool.poetry]
name = "runner"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
boto3 = "^1.11"
msrest = "^0.6.11"
azure-cognitiveservices-search-imagesearch = "^2.0"
azure-cognitiveservices-search-websearch = "^2.0"
bs4 = "^0.0.1"
flask = "^1.1"
flask-cors = "^3.0"
textblob = "^0.15.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
